refactor: 예약 가능 시간 결과 모델 위치 변경
